轻微改动: GameActivity.java 修复车辆没名字的Bug 发现InforPoolClient.java有Bug,在380行,程序运行一段时间后会出现空指针,未找到原因

+++ XRace_V2/src/com/sa/xrace/client/pool/CarInforClient.java
@@ -60,7 +60,7 @@ public final class CarInforClient {
     private float nAccidenceDirection;
     private float nAccidenceChangeDirection;
 
-    private short timeDelay;// time delay of this car in network transmit
+    private int timeDelay;// time delay of this car in network transmit
 
     private int carListName;
 
@@ -278,11 +278,11 @@ public final class CarInforClient {
         nAccidenceDirection = accidenceDirection;
     }
 
-    public short getTimeDelay() {
+    public int getTimeDelay() {
         return timeDelay;
     }
 
-    public void setTimeDelay(short timeDelay) {
+    public void setTimeDelay(int timeDelay) {
         this.timeDelay = timeDelay;
     }
